% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{PO.EN.predict}
\alias{PO.EN.predict}
\title{PO-EN predicting function}
\usage{
PO.EN.predict(X, beta)
}
\arguments{
\item{X}{Input design matrix. Should not include the intercept vector.}

\item{beta}{A coefficients vector from the PO-EN fitting function.}
}
\description{
A prediction function using the linear predictor of PO-EN fitting results.
}
\examples{
PO.EN.cv<-cv.PO.EN(x_train,y_train,input.pi=seq(0.01,0.4,length.out=10))
PO.EN.beta<-PO.EN(x_train,y_train,lambda=PO.EN.cv$lambda.min,
           true.prob=PO.EN.cv$pi,beta_start=rep(0,ncol(x_train)+1))
predictions<-PO.EN.predict(x_test,PO.EN.beta)
roc(y_test~predictions)
}
