% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{PO.EN.predict}
\alias{PO.EN.predict}
\title{PO-EN predicting function}
\usage{
PO.EN.predict(X, beta)
}
\arguments{
\item{X}{Input design matrix. Should not include the intercept vector.}

\item{beta}{A coefficients vector from the PO-EN fitting function.}
}
\description{
A prediction function using the linear predictor of PO-EN fitting results.
}
